@@ -1,7 +1,7 @@
 // coverage:ignore-file
 // GENERATED CODE - DO NOT MODIFY BY HAND
 // ignore_for_file: type=lint
-// ignore_for_file: unused_element, deprecated_member_use, deprecated_member_use_from_same_package, use_function_type_syntax_for_parameters, unnecessary_const, avoid_init_to_null, invalid_override_different_default_values_named, prefer_expression_function_bodies, annotate_overrides, invalid_annotation_target
+// ignore_for_file: unused_element, deprecated_member_use, deprecated_member_use_from_same_package, use_function_type_syntax_for_parameters, unnecessary_const, avoid_init_to_null, invalid_override_different_default_values_named, prefer_expression_function_bodies, annotate_overrides, invalid_annotation_target, unnecessary_question_mark
 
 part of 'auth_tokens.dart';
 
@@ -30,7 +30,8 @@ mixin _$AuthTokens {
 abstract class $AuthTokensCopyWith<$Res> {
   factory $AuthTokensCopyWith(
           AuthTokens value, $Res Function(AuthTokens) then) =
-      _$AuthTokensCopyWithImpl<$Res>;
+      _$AuthTokensCopyWithImpl<$Res, AuthTokens>;
+  @useResult
   $Res call(
       {String accessToken,
       String refreshToken,
@@ -39,47 +40,52 @@ abstract class $AuthTokensCopyWith<$Res> {
 }
 
 /// @nodoc
-class _$AuthTokensCopyWithImpl<$Res> implements $AuthTokensCopyWith<$Res> {
+class _$AuthTokensCopyWithImpl<$Res, $Val extends AuthTokens>
+    implements $AuthTokensCopyWith<$Res> {
   _$AuthTokensCopyWithImpl(this._value, this._then);
 
-  final AuthTokens _value;
   // ignore: unused_field
-  final $Res Function(AuthTokens) _then;
+  final $Val _value;
+  // ignore: unused_field
+  final $Res Function($Val) _then;
 
+  @pragma('vm:prefer-inline')
   @override
   $Res call({
-    Object? accessToken = freezed,
-    Object? refreshToken = freezed,
-    Object? tokenType = freezed,
-    Object? expiresIn = freezed,
+    Object? accessToken = null,
+    Object? refreshToken = null,
+    Object? tokenType = null,
+    Object? expiresIn = null,
   }) {
     return _then(_value.copyWith(
-      accessToken: accessToken == freezed
+      accessToken: null == accessToken
           ? _value.accessToken
           : accessToken // ignore: cast_nullable_to_non_nullable
               as String,
-      refreshToken: refreshToken == freezed
+      refreshToken: null == refreshToken
           ? _value.refreshToken
           : refreshToken // ignore: cast_nullable_to_non_nullable
               as String,
-      tokenType: tokenType == freezed
+      tokenType: null == tokenType
           ? _value.tokenType
           : tokenType // ignore: cast_nullable_to_non_nullable
               as String,
-      expiresIn: expiresIn == freezed
+      expiresIn: null == expiresIn
           ? _value.expiresIn
           : expiresIn // ignore: cast_nullable_to_non_nullable
               as double,
-    ));
+    ) as $Val);
   }
 }
 
 /// @nodoc
-abstract class _$AuthTokensCopyWith<$Res> implements $AuthTokensCopyWith<$Res> {
-  factory _$AuthTokensCopyWith(
-          _AuthTokens value, $Res Function(_AuthTokens) then) =
-      __$AuthTokensCopyWithImpl<$Res>;
+abstract class _$$_AuthTokensCopyWith<$Res>
+    implements $AuthTokensCopyWith<$Res> {
+  factory _$$_AuthTokensCopyWith(
+          _$_AuthTokens value, $Res Function(_$_AuthTokens) then) =
+      __$$_AuthTokensCopyWithImpl<$Res>;
   @override
+  @useResult
   $Res call(
       {String accessToken,
       String refreshToken,
@@ -88,36 +94,35 @@ abstract class _$AuthTokensCopyWith<$Res> implements $AuthTokensCopyWith<$Res> {
 }
 
 /// @nodoc
-class __$AuthTokensCopyWithImpl<$Res> extends _$AuthTokensCopyWithImpl<$Res>
-    implements _$AuthTokensCopyWith<$Res> {
-  __$AuthTokensCopyWithImpl(
-      _AuthTokens _value, $Res Function(_AuthTokens) _then)
-      : super(_value, (v) => _then(v as _AuthTokens));
+class __$$_AuthTokensCopyWithImpl<$Res>
+    extends _$AuthTokensCopyWithImpl<$Res, _$_AuthTokens>
+    implements _$$_AuthTokensCopyWith<$Res> {
+  __$$_AuthTokensCopyWithImpl(
+      _$_AuthTokens _value, $Res Function(_$_AuthTokens) _then)
+      : super(_value, _then);
 
-  @override
-  _AuthTokens get _value => super._value as _AuthTokens;
-
+  @pragma('vm:prefer-inline')
   @override
   $Res call({
-    Object? accessToken = freezed,
-    Object? refreshToken = freezed,
-    Object? tokenType = freezed,
-    Object? expiresIn = freezed,
+    Object? accessToken = null,
+    Object? refreshToken = null,
+    Object? tokenType = null,
+    Object? expiresIn = null,
   }) {
-    return _then(_AuthTokens(
-      accessToken: accessToken == freezed
+    return _then(_$_AuthTokens(
+      accessToken: null == accessToken
           ? _value.accessToken
           : accessToken // ignore: cast_nullable_to_non_nullable
               as String,
-      refreshToken: refreshToken == freezed
+      refreshToken: null == refreshToken
           ? _value.refreshToken
           : refreshToken // ignore: cast_nullable_to_non_nullable
               as String,
-      tokenType: tokenType == freezed
+      tokenType: null == tokenType
           ? _value.tokenType
           : tokenType // ignore: cast_nullable_to_non_nullable
               as String,
-      expiresIn: expiresIn == freezed
+      expiresIn: null == expiresIn
           ? _value.expiresIn
           : expiresIn // ignore: cast_nullable_to_non_nullable
               as double,
@@ -153,27 +158,26 @@ class _$_AuthTokens extends _AuthTokens {
   bool operator ==(dynamic other) {
     return identical(this, other) ||
         (other.runtimeType == runtimeType &&
-            other is _AuthTokens &&
-            const DeepCollectionEquality()
-                .equals(other.accessToken, accessToken) &&
-            const DeepCollectionEquality()
-                .equals(other.refreshToken, refreshToken) &&
-            const DeepCollectionEquality().equals(other.tokenType, tokenType) &&
-            const DeepCollectionEquality().equals(other.expiresIn, expiresIn));
+            other is _$_AuthTokens &&
+            (identical(other.accessToken, accessToken) ||
+                other.accessToken == accessToken) &&
+            (identical(other.refreshToken, refreshToken) ||
+                other.refreshToken == refreshToken) &&
+            (identical(other.tokenType, tokenType) ||
+                other.tokenType == tokenType) &&
+            (identical(other.expiresIn, expiresIn) ||
+                other.expiresIn == expiresIn));
   }
 
   @override
-  int get hashCode => Object.hash(
-      runtimeType,
-      const DeepCollectionEquality().hash(accessToken),
-      const DeepCollectionEquality().hash(refreshToken),
-      const DeepCollectionEquality().hash(tokenType),
-      const DeepCollectionEquality().hash(expiresIn));
+  int get hashCode =>
+      Object.hash(runtimeType, accessToken, refreshToken, tokenType, expiresIn);
 
   @JsonKey(ignore: true)
   @override
-  _$AuthTokensCopyWith<_AuthTokens> get copyWith =>
-      __$AuthTokensCopyWithImpl<_AuthTokens>(this, _$identity);
+  @pragma('vm:prefer-inline')
+  _$$_AuthTokensCopyWith<_$_AuthTokens> get copyWith =>
+      __$$_AuthTokensCopyWithImpl<_$_AuthTokens>(this, _$identity);
 }
 
 abstract class _AuthTokens extends AuthTokens {
@@ -185,15 +189,15 @@ abstract class _AuthTokens extends AuthTokens {
   const _AuthTokens._() : super._();
 
   @override
-  String get accessToken => throw _privateConstructorUsedError;
+  String get accessToken;
   @override
-  String get refreshToken => throw _privateConstructorUsedError;
+  String get refreshToken;
   @override
-  String get tokenType => throw _privateConstructorUsedError;
+  String get tokenType;
   @override
-  double get expiresIn => throw _privateConstructorUsedError;
+  double get expiresIn;
   @override
   @JsonKey(ignore: true)
-  _$AuthTokensCopyWith<_AuthTokens> get copyWith =>
+  _$$_AuthTokensCopyWith<_$_AuthTokens> get copyWith =>
       throw _privateConstructorUsedError;
 }
