@@ -1,11 +1,20 @@
 import 'package:auto_route/auto_route.dart';
-import 'package:flutter_template/presentation/feature/home/home_screen.dart';
-import 'package:flutter_template/presentation/feature/splash/splash_screen.dart';
+import 'package:flutter_template/presentation/routes/router.gr.dart';
 
-@MaterialAutoRouter(
-  routes: <AutoRoute>[
-    AutoRoute(page: SplashScreen, path: '/'),
-    AutoRoute(page: HomeScreen, path: '/home'),
-  ],
-)
-class $AppRouter {}
+@AutoRouterConfig()
+class AppRouter extends $AppRouter {
+  @override
+  RouteType get defaultRouteType => const RouteType.material();
+
+  @override
+  List<AutoRoute> get routes => [
+        AutoRoute(
+          page: SplashRoute.page,
+          path: '/',
+        ),
+        AutoRoute(
+          page: HomeRoute.page,
+          path: '/home',
+        ),
+      ];
+}
